[Refactor] 공통 로그아웃 로직을 useLogout 훅으로 통합
